Solution for problem 2
Worked on a solution with no built in methods but found it too hard so
worked on a solution with built in methods and then put an answer with
just one line of code. Can come back to this to find out how to use just one method call
to replaceAll instead of two calls once I have finished working on the
problems.